Perl Serial Test
First release to test communication with Quantar. Code seems to be ok, but I am loosing a random byte every now and then when voice is Rx. Tested on Raspberry Pi and MacBook Pro with the same results. Using the same Mac hardware with VB6 or GoLang none of the bytes is lost, so it should be a bug related to Perl.